board.linux: make linux login more robust
There is already the login_delay parameter, which
helps to make login more robust if console is
flooded on boot up time.

tbot waits for the string "login: " and expects that
there are no more characters after it. But in case the
console get flooded with characters on boot up time,
it can happen that there are characters in the same line
after "login: ".

Currently, read_until_prompt waits endless, when trying
to get "login: " and so tbot never login into linux.

rework this part, so read_until_prompt waits "login_delay"
seconds, and after that, trigger a new login prompt.

Signed-off-by: Heiko Schocher <hs@denx.de>
Series-to: tbot@lists.denx.de